+++ lib/runners/crosvm.nix
@@ -1,6 +1,7 @@
 { pkgs
 , microvmConfig
 , macvtapFds
+, linuxTarget
 , ...
 }:
 
@@ -17,7 +18,7 @@ let
 
   kernelPath = {
     x86_64-linux = "${kernel.dev}/vmlinux";
-    aarch64-linux = "${kernel.out}/${pkgs.stdenv.hostPlatform.linux-kernel.target}";
+    aarch64-linux = "${kernel.out}/${linuxTarget}";
   }.${system};
 
   gpuParams = {